src/vapiserve/integrations/storage/aws_s3.py
"""AWS S3 storage integration."""

import os
from typing import Any, Dict, List, Optional, Union, BinaryIO

try:
    import boto3
    from botocore.exceptions import ClientError
    HAS_BOTO = True
except ImportError:
    HAS_BOTO = False

from .base import StorageProvider


class AWSS3Provider(StorageProvider):
    """AWS S3 storage integration."""
    
    def __init__(
        self,
        bucket_name: str,
        aws_access_key_id: Optional[str] = None,
        aws_secret_access_key: Optional[str] = None,
        aws_session_token: Optional[str] = None,
        region_name: Optional[str] = None,
        endpoint_url: Optional[str] = None,
    ):
        """Initialize the AWS S3 storage provider.
        
        Args:
            bucket_name: Name of the S3 bucket
            aws_access_key_id: AWS access key ID
            aws_secret_access_key: AWS secret access key
            aws_session_token: AWS session token
            region_name: AWS region name
            endpoint_url: Custom endpoint URL for S3-compatible storage
        """
        if not HAS_BOTO:
            raise ImportError(
                "AWS S3 dependencies not installed. "
                "Install with 'pip install boto3'"
            )
        
        self.bucket_name = bucket_name
        
        # Use provided credentials or fall back to environment variables/IAM role
        self.aws_access_key_id = aws_access_key_id or os.environ.get("AWS_ACCESS_KEY_ID")
        self.aws_secret_access_key = aws_secret_access_key or os.environ.get("AWS_SECRET_ACCESS_KEY")
        self.aws_session_token = aws_session_token or os.environ.get("AWS_SESSION_TOKEN")
        self.region_name = region_name or os.environ.get("AWS_REGION")
        self.endpoint_url = endpoint_url
        
        # Initialize S3 client
        self.client = boto3.client(
            "s3",
            aws_access_key_id=self.aws_access_key_id,
            aws_secret_access_key=self.aws_secret_access_key,
            aws_session_token=self.aws_session_token,
            region_name=self.region_name,
            endpoint_url=self.endpoint_url
        )
        
        # Initialize S3 resource for higher-level operations
        self.resource = boto3.resource(
            "s3",
            aws_access_key_id=self.aws_access_key_id,
            aws_secret_access_key=self.aws_secret_access_key,
            aws_session_token=self.aws_session_token,
            region_name=self.region_name,
            endpoint_url=self.endpoint_url
        )
        
        # Get the bucket
        self.bucket = self.resource.Bucket(self.bucket_name)
    
    def upload_file(
        self, 
        file_content: Union[str, bytes, BinaryIO],
        destination_path: str,
        content_type: Optional[str] = None,
        metadata: Optional[Dict[str, str]] = None,
        **kwargs
    ) -> Dict[str, Any]:
        """Upload a file to AWS S3.
        
        Args:
            file_content: Content of the file to upload
            destination_path: Path in the bucket where the file should be stored
            content_type: MIME type of the file
            metadata: Additional metadata to store with the file
            **kwargs: Additional parameters like ACL, StorageClass, etc.
            
        Returns:
            Dict containing the upload result
        """
        # TODO: Implement file upload using boto3
        # Handle different types of file_content (str, bytes, file-like object)
        # Set content_type and metadata
        # Handle extra S3-specific parameters in kwargs
        
        raise NotImplementedError("AWS S3 file upload not implemented yet")
    
    def download_file(
        self,
        file_path: str,
        **kwargs
    ) -> bytes:
        """Download a file from AWS S3.
        
        Args:
            file_path: Path of the file in the bucket
            **kwargs: Additional parameters
            
        Returns:
            File content as bytes
        """
        # TODO: Implement file download using boto3
        
        raise NotImplementedError("AWS S3 file download not implemented yet")
    
    def get_file_metadata(
        self,
        file_path: str,
        **kwargs
    ) -> Dict[str, Any]:
        """Get metadata for a file in AWS S3.
        
        Args:
            file_path: Path of the file in the bucket
            **kwargs: Additional parameters
            
        Returns:
            Dict containing file metadata
        """
        # TODO: Implement file metadata retrieval using boto3
        
        raise NotImplementedError("AWS S3 file metadata retrieval not implemented yet")
    
    def delete_file(
        self,
        file_path: str,
        **kwargs
    ) -> bool:
        """Delete a file from AWS S3.
        
        Args:
            file_path: Path of the file in the bucket to delete
            **kwargs: Additional parameters
            
        Returns:
            True if deletion was successful, False otherwise
        """
        # TODO: Implement file deletion using boto3
        
        raise NotImplementedError("AWS S3 file deletion not implemented yet")
    
    def list_files(
        self,
        directory_path: str,
        recursive: bool = False,
        **kwargs
    ) -> List[Dict[str, Any]]:
        """List files in a directory in AWS S3.
        
        Args:
            directory_path: Path of the directory in the bucket to list
            recursive: Whether to list files in subdirectories
            **kwargs: Additional parameters like MaxKeys, StartAfter, etc.
            
        Returns:
            List of file metadata
        """
        # TODO: Implement file listing using boto3
        
        raise NotImplementedError("AWS S3 file listing not implemented yet")
    
    def generate_signed_url(
        self,
        file_path: str,
        expiration: int,
        **kwargs
    ) -> str:
        """Generate a signed URL for temporary access to a file in AWS S3.
        
        Args:
            file_path: Path of the file in the bucket
            expiration: Expiration time in seconds
            **kwargs: Additional parameters
            
        Returns:
            Signed URL
        """
        # TODO: Implement signed URL generation using boto3
        
        raise NotImplementedError("AWS S3 signed URL generation not implemented yet") 