src/vapiserve/integrations/storage/google_cloud_storage.py
"""Google Cloud Storage integration."""

import os
import io
from typing import Any, Dict, List, Optional, Union, BinaryIO

try:
    from google.cloud import storage
    from google.oauth2 import service_account
    HAS_GCS = True
except ImportError:
    HAS_GCS = False

from .base import StorageProvider


class GoogleCloudStorageProvider(StorageProvider):
    """Google Cloud Storage integration."""
    
    def __init__(
        self,
        bucket_name: str,
        credentials_json_or_path: Optional[str] = None,
        project_id: Optional[str] = None,
    ):
        """Initialize the Google Cloud Storage provider.
        
        Args:
            bucket_name: Name of the GCS bucket
            credentials_json_or_path: JSON string or path to service account credentials
            project_id: Google Cloud project ID
        """
        if not HAS_GCS:
            raise ImportError(
                "Google Cloud Storage dependencies not installed. "
                "Install with 'pip install google-cloud-storage'"
            )
        
        self.bucket_name = bucket_name
        self.project_id = project_id or os.environ.get("GOOGLE_CLOUD_PROJECT")
        
        # Initialize client with credentials
        if credentials_json_or_path:
            # Check if it's a file path
            if os.path.exists(credentials_json_or_path):
                # Load credentials from file
                self.credentials = service_account.Credentials.from_service_account_file(
                    credentials_json_or_path
                )
            else:
                # Assume it's a JSON string
                # TODO: Implement loading credentials from JSON string
                raise NotImplementedError("Loading credentials from JSON string not implemented yet")
        else:
            # Use default credentials
            self.credentials = None
        
        # Initialize storage client
        self.client = storage.Client(
            credentials=self.credentials,
            project=self.project_id
        )
        
        # Get the bucket
        self.bucket = self.client.bucket(self.bucket_name)
    
    def upload_file(
        self, 
        file_content: Union[str, bytes, BinaryIO],
        destination_path: str,
        content_type: Optional[str] = None,
        metadata: Optional[Dict[str, str]] = None,
        **kwargs
    ) -> Dict[str, Any]:
        """Upload a file to Google Cloud Storage.
        
        Args:
            file_content: Content of the file to upload
            destination_path: Path in the bucket where the file should be stored
            content_type: MIME type of the file
            metadata: Additional metadata to store with the file
            **kwargs: Additional parameters
            
        Returns:
            Dict containing the upload result
        """
        # TODO: Implement file upload using Google Cloud Storage API
        
        raise NotImplementedError("Google Cloud Storage file upload not implemented yet")
    
    def download_file(
        self,
        file_path: str,
        **kwargs
    ) -> bytes:
        """Download a file from Google Cloud Storage.
        
        Args:
            file_path: Path of the file in the bucket
            **kwargs: Additional parameters
            
        Returns:
            File content as bytes
        """
        # TODO: Implement file download using Google Cloud Storage API
        
        raise NotImplementedError("Google Cloud Storage file download not implemented yet")
    
    def get_file_metadata(
        self,
        file_path: str,
        **kwargs
    ) -> Dict[str, Any]:
        """Get metadata for a file in Google Cloud Storage.
        
        Args:
            file_path: Path of the file in the bucket
            **kwargs: Additional parameters
            
        Returns:
            Dict containing file metadata
        """
        # TODO: Implement file metadata retrieval using Google Cloud Storage API
        
        raise NotImplementedError("Google Cloud Storage file metadata retrieval not implemented yet")
    
    def delete_file(
        self,
        file_path: str,
        **kwargs
    ) -> bool:
        """Delete a file from Google Cloud Storage.
        
        Args:
            file_path: Path of the file in the bucket to delete
            **kwargs: Additional parameters
            
        Returns:
            True if deletion was successful, False otherwise
        """
        # TODO: Implement file deletion using Google Cloud Storage API
        
        raise NotImplementedError("Google Cloud Storage file deletion not implemented yet")
    
    def list_files(
        self,
        directory_path: str,
        recursive: bool = False,
        **kwargs
    ) -> List[Dict[str, Any]]:
        """List files in a directory in Google Cloud Storage.
        
        Args:
            directory_path: Path of the directory in the bucket to list
            recursive: Whether to list files in subdirectories
            **kwargs: Additional parameters
            
        Returns:
            List of file metadata
        """
        # TODO: Implement file listing using Google Cloud Storage API
        
        raise NotImplementedError("Google Cloud Storage file listing not implemented yet")
    
    def generate_signed_url(
        self,
        file_path: str,
        expiration: int,
        **kwargs
    ) -> str:
        """Generate a signed URL for temporary access to a file in Google Cloud Storage.
        
        Args:
            file_path: Path of the file in the bucket
            expiration: Expiration time in seconds
            **kwargs: Additional parameters
            
        Returns:
            Signed URL
        """
        # TODO: Implement signed URL generation using Google Cloud Storage API
        
        raise NotImplementedError("Google Cloud Storage signed URL generation not implemented yet") 