pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vapiserve"
version = "0.0.1"
description = "A framework for building API servers for Vapi custom tools with integrated service providers"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
    { name = "Mahimai Raja", email = "mahimairaja3@gmail.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.100.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.22.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pytest",
    "pytest-cov",
    "mypy",
    "ruff",
]
scheduling = [
    "google-auth>=2.22.0",
    "google-api-python-client>=2.97.0",
    "microsoft-graph-core>=0.0.1",
    "msal>=1.20.0",
]
tasks = [
    "todoist-api-python>=2.0.0",
]
communication = [
    "slack-sdk>=3.21.3",
    "twilio>=8.5.0",
]
storage = [
    "boto3>=1.26.0",
    "google-cloud-storage>=2.8.0",
]
email = [
    "sendgrid>=6.0.0",
]
ai = [
    "openai>=0.27.8",
    "anthropic>=0.5.0",
]
all-integrations = [
    "vapiserve[scheduling,tasks,communication,storage,email,ai]",
]
ngrok = [
    "pyngrok>=5.2.1",
]

[project.urls]
"Homepage" = "https://github.com/vapi-ai/vapiserve"
"Bug Tracker" = "https://github.com/vapi-ai/vapiserve/issues"
"Documentation" = "https://vapiserve.readthedocs.io/"

[project.scripts]
vapiserve = "vapiserve.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/vapiserve"]

[tool.hatch.build.targets.sdist]
include = [
    "src/vapiserve",
    "LICENSE",
    "README.md",
    "pyproject.toml",
    "cookbook",
] 